+++ lib/ace/commands/default_commands.js
@@ -45,7 +45,7 @@ exports.commands = [{
     name: "showKeyboardShortcuts",
     bindKey: bindKey("Ctrl-Alt-H", "Command-Alt-H"),
     exec: function (editor) {
-        config.loadModule("ace/ext/showKeyboardShortcuts", function (e) {
+        config.loadModule("ace/ext/show_keyboard_shortcuts", function (e) {
             e(editor);
         });
     },
